turn current round to game-wide data using SO

+++ ScriptableObjects/Variables/IntVariable.cs
@@ -21,7 +21,7 @@ public class IntVariable : ScriptableObject
         onValueChanged.RaiseEvent(Value);
     }
 
-    public float GetValue() { return Value; }
+    public int GetValue() { return Value; }
 
     public IntEventChannelSO onValueChanged;
 }
